Settings: Move Connectivity down and give it an icon (like on android and ios)
