+++ images/licenses
@@ -7,5 +7,6 @@ Google's material icons - Apache 2:
 - icons/favorite.svg
 - icons/open_in_new.svg
 - icons/schedule.svg
+- icons/swap_vert.svg
 
 // TODO
